(core) simplify the Wire trait implementation
This commit doesn't really simplify aynthing to be honest. It adds a lot
of boilerplate code that just add to the complexity. My goal was to have
a trait called `WireDynamic` that array and string types implement.
Then, we could have a blanket implementation of `Wire`. However, the
compiler throws an error since we have a blanket implementation in terms
of `WirePrimitive` also. We can't have multiple blanker implementations
for the same trait, because the compiler cannot guarantee that
`WirePrimitive` and `WireDynamic` are not ever implemented by the same
type. I quite liked how I had written `WireDynamic`, so I didn't want to
remove it. This is why the complexity.

But, I learned something. So, it is a win I guess.