(core) add a `Cow`-like type `Fd` to represent file descriptors
The issue:

The type of the file descriptor changes depending on whether we are
receiving the message or sending it. We use `OwnedFd` to receive and
`BorrowedFd<'fd>` to send. This means we cannot have a single type that
works for both servers and clients.

Possible solutions:

- Do what `wayland-rs` does and generate separate modules named `server`
  and `client`. In each module, generate the correct definition. Doing
  this has no runtime cost, as the type is already known at compile
  time.
- Have a single type with a generic `Fd` parameter. Set it to
  `BorrowedFd<'fd>` if you are sending the message, and set it to
  `OwnedFd` if you are receiving it. This is equivalent to the previous
  approach in the sense that the compiler essentially generates two
  types for both use cases. This is the approach we currently use.

My solution:

We use `Cow` for strings and arrays. This allows us to use a `String` if
we are receiving the message, and use a `&'buf str` if we are sending
it. I use this idea to create a new enum type called `Fd` that has two
variants: `Borrowed(BorrowedFd<'fd>)` and `Owned(OwnedFd)`. This allows
us to use a single type for all use cases. This has the disadvantage of
having to match on all the variants even when we know the type in
compile time (which is almost always). I decided to use it to make the
code generation simpler. The approach may be changed in the future
depending on further usage.